requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
facebook-sdk==3.1.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
jupyter-client==8.3.0
jupyter-core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
tqdm==4.66.0
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
zipp==3.16.2
